requirements.txt
asgiref==3.8.1
distlib==0.3.8
Django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
filelock==3.14.0
gunicorn==22.0.0
packaging==24.1
platformdirs==4.2.1
PyJWT==2.8.0
sqlparse==0.5.0
tzdata==2024.1
virtualenv==20.26.1
virtualenvwrapper-win==1.2.7
